Handle nodes array in traversal to config->nodes.
When a traversal has no nodes specified it will get all nodes, which
is achieved by referencing the config->nodes array. However,
the nodes array is cleaned in several functions when arrays are
flattened. So, now we check if the traversal array is the
config->nodes array before doing the flattening and cleaning