Insert the refactor and name change.
This commit transforms it a bit more as traversals and forms th
e base for the bootstrapping. Hopefully the mapping will be more 1:1
It also removes some things from palm and integrates some other stuff
reducing the overall runtime of coconut.
This is now rather minimal, it almost implements everything from the
older dynamic backend. But it is not complete yet, the complete one
will come in the bootstrapped version.